added method for admins to make store transactions

+++ src/main/java/com/mycompany/e_commerce/service/UserService.java
@@ -19,6 +19,7 @@ import com.mycompany.e_commerce.entity.User;
 import com.mycompany.e_commerce.entity.enums.Role;
 import com.mycompany.e_commerce.exception.customexception.CustomeException;
 import com.mycompany.e_commerce.exception.customexception.TokenInvalidException;
+import com.mycompany.e_commerce.exception.customexception.UnauthorizedAccessException;
 import com.mycompany.e_commerce.exception.customexception.UserAlreadyExistException;
 import com.mycompany.e_commerce.exception.customexception.UserNameOrPasswordWrongException;
 import com.mycompany.e_commerce.exception.customexception.UserNotFoundException;
@@ -101,8 +102,7 @@ public class UserService {
 
     public String login(LoginRequest login) {
         User user = userRepository.findByUserNameAndPassword(login.getUserName(), login.getPassword()).orElseThrow(
-            () -> new UserNameOrPasswordWrongException(CustomeException.USER_NAME_OR_PASSWORD_IS_WRONG)
-        );
+                () -> new UserNameOrPasswordWrongException(CustomeException.USER_NAME_OR_PASSWORD_IS_WRONG));
         return jwtManager.generateToken(user.getId());
     }
 
@@ -113,7 +113,8 @@ public class UserService {
     }
 
     protected User getUserById(Long id) {
-        return userRepository.findById(id).orElseThrow(() -> new UserNotFoundException(CustomeException.USER_NOT_FOUND));
+        return userRepository.findById(id)
+                .orElseThrow(() -> new UserNotFoundException(CustomeException.USER_NOT_FOUND));
     }
 
     protected List<User> getAllUsers() {
@@ -121,42 +122,56 @@ public class UserService {
     }
 
     @Transactional
-    public void updateCustomer(CustomerUpdateRequest customer) {
+    public void updateCustomer(String token, CustomerUpdateRequest customer) {
+        Long id = getUserIdFromToken(token);
         User currentUser = getUserById(customer.getId());
-        currentUser.setName(customer.getName());
-        currentUser.setUserName(customer.getUserName());
-        currentUser.setPassword(customer.getPassword());
-        currentUser.setUpdateAt(
-                LocalDateTime.now().toString());
-        userRepository.save(currentUser);
+        if (id == currentUser.getId()) {
+            currentUser.setEmail(customer.getEmail());
+            currentUser.setUserName(customer.getUserName());
+            currentUser.setPassword(customer.getPassword());
+            currentUser.setPhone(customer.getPhone());
+            currentUser.setAddress(customer.getAddress());
+            currentUser.setUpdateAt(
+                    LocalDateTime.now().toString());
+            userRepository.save(currentUser);
+        } else {
+            throw new UnauthorizedAccessException(CustomeException.UNAUTHORIZED_ACCESS);
+        }
     }
 
     @Transactional
-    public void updateAdmin(AdminUpdateRequest admin) {
+    public void updateAdmin(String token, AdminUpdateRequest admin) {
+        Long id = getUserIdFromToken(token);
         User currentUser = getUserById(admin.getId());
-        currentUser.setEmail(admin.getEmail());
-        currentUser.setUserName(admin.getUserName());
-        currentUser.setPassword(admin.getPassword());
-        currentUser.setPhone(admin.getPhone());
-        currentUser.setUpdateAt(
-                LocalDateTime.now().toString());
-        userRepository.save(currentUser);
+        if (id == currentUser.getId()) {
+            currentUser.setEmail(admin.getEmail());
+            currentUser.setUserName(admin.getUserName());
+            currentUser.setPassword(admin.getPassword());
+            currentUser.setPhone(admin.getPhone());
+            currentUser.setUpdateAt(
+                    LocalDateTime.now().toString());
+            userRepository.save(currentUser);
+        } else {
+            throw new UnauthorizedAccessException(CustomeException.UNAUTHORIZED_ACCESS);
+        }
     }
 
     @Transactional
     public void deleteAdmin(String token, Long id) {
         User user = getUserById(getUserIdFromToken(token));
-        if (user.getRole().equals(Role.ADMIN.toString())) {
+        if (user.getRole().equals(Role.ADMIN.toString()) && !user.getId().equals(id)) {
             userRepository.deleteById(id);
+        } else {
+            throw new UnauthorizedAccessException(CustomeException.UNAUTHORIZED_ACCESS);
         }
     }
 
     public Double getTotalSalesForToday(String token, Long storeId) {
         User user = getUserById(getUserIdFromToken(token));
         if (user.getRole().equals(Role.ADMIN.toString())) {
-            return orderService.getTotalSalesForToday(storeId);            
+            return orderService.getTotalSalesForToday(storeId);
         } else {
-            return 0.0;     
+            return 0.0;
         }
     }
 
@@ -170,6 +185,7 @@ public class UserService {
     }
 
     private Long getUserIdFromToken(String token) {
-        return jwtManager.validToken(token).orElseThrow(() -> new TokenInvalidException(CustomeException.TOKEN_INVALID));
+        return jwtManager.validToken(token)
+                .orElseThrow(() -> new TokenInvalidException(CustomeException.TOKEN_INVALID));
     }
 }
